"GTA3" autosplitter: added resetting stuff
Luckily the GameState numbers are identical to Vice City, and the way we
check for the starting split means we don't need to check if a save is
being loaded or not (although that needs some testing). Also need to
check if the offset works with the GameState address I found.