"GTA3" autosplitter: changed version detection
I only now realised the memory addresses for version 1.0/1.1 were
written wrong, so I fixed that and do the check in a different way now
(not sure if it's better though). Tested 1.1 and it works fine.

I still don't think this is the best way to detect versions; might move
to using ModuleMemorySize for everything at some point; sometimes the
autosplitter just can't detect the game for example.